Restructure the Layout and fix styling issues
